refactor(trader): v1.9 share-based stake + chainlink pre-close outcome
Two related changes:

1. **Share-based Martingale ladder** (v1.8.2):
   - LadderState: `base_usd: Decimal` → `base_shares: u32`
   - WindowOutcome::Won gains `cost_usd` field (was implicit via current_bet)
   - 5-share floor consistent with Polymarket CLOB minimum
   - Doubling sequence in shares: 5/10/20/40/80/...

2. **Chainlink-based outcome at t=window_close-4s** (v1.9 hold mode):
   - New `--exit-rule hold` behavior: query Chainlink BTC/USD on Polygon,
     compare to `price_to_beat`, emit Won/Lost at t=296s (4s buffer before
     close). Skips sell — Polymarket Auto-Redeem credits winning shares.
   - Preserves scheduler throughput (returns ~4s before next boundary).
   - Strict Martingale: each window resolves before next entry, no lag.
   - Chainlink RPC failure falls back to gamma resolver (may miss next
     window but gets correct outcome).
   - 358 lib tests pass.

+++ src/app.rs
@@ -291,7 +291,7 @@ mod tests {
             ts: Utc::now(),
             session_id: Uuid::nil(),
             kind: TraderEventKind::SessionStarted,
-            ladder: LadderState::new(Direction::Up, Decimal::from(5), 5, Utc::now()),
+            ladder: LadderState::new(Direction::Up, 5, 5, Utc::now()),
         };
         handle_event(&mut s, AppEvent::TraderEvent(ev.clone()), &tx);
         assert_eq!(s.trader_log.len(), 1);
@@ -311,7 +311,7 @@ mod tests {
             let ev = TraderEvent {
                 ts: Utc::now(), session_id: Uuid::nil(),
                 kind: TraderEventKind::SessionStarted,
-                ladder: LadderState::new(Direction::Up, Decimal::from(5), 5, Utc::now()),
+                ladder: LadderState::new(Direction::Up, 5, 5, Utc::now()),
             };
             handle_event(&mut s, AppEvent::TraderEvent(ev), &tx);
         }
@@ -337,7 +337,7 @@ mod tests {
         let ev = TraderEvent {
             ts: now - Cd::seconds(120), session_id: Uuid::nil(),
             kind: TraderEventKind::SessionStarted,
-            ladder: LadderState::new(Direction::Up, Decimal::from(5), 5, now),
+            ladder: LadderState::new(Direction::Up, 5, 5, now),
         };
         assert_eq!(compute_trader_health(&Some(ev), now), TraderHealth::Healthy);
     }
@@ -354,7 +354,7 @@ mod tests {
         let ev = TraderEvent {
             ts: now - Cd::seconds(8 * 60), session_id: Uuid::nil(),
             kind: TraderEventKind::SessionStarted,
-            ladder: LadderState::new(Direction::Up, Decimal::from(5), 5, now),
+            ladder: LadderState::new(Direction::Up, 5, 5, now),
         };
         assert_eq!(compute_trader_health(&Some(ev), now), TraderHealth::Lagging);
     }
@@ -371,7 +371,7 @@ mod tests {
         let ev = TraderEvent {
             ts: now - Cd::seconds(15 * 60), session_id: Uuid::nil(),
             kind: TraderEventKind::SessionStarted,
-            ladder: LadderState::new(Direction::Up, Decimal::from(5), 5, now),
+            ladder: LadderState::new(Direction::Up, 5, 5, now),
         };
         assert_eq!(compute_trader_health(&Some(ev), now), TraderHealth::Stale);
     }
@@ -388,7 +388,7 @@ mod tests {
         let ev = TraderEvent {
             ts: now - Cd::seconds(30), session_id: Uuid::nil(),
             kind: TraderEventKind::SessionStopped { reason: StopReason::CapReached },
-            ladder: LadderState::new(Direction::Up, Decimal::from(5), 5, now),
+            ladder: LadderState::new(Direction::Up, 5, 5, now),
         };
         assert_eq!(compute_trader_health(&Some(ev), now), TraderHealth::Stopped);
     }
@@ -423,7 +423,7 @@ mod tests {
         let (mins_tx, mut mins_rx) = mpsc::channel::<u32>(8);
         state.window_minutes_tx = Some(mins_tx);
 
-        let ladder = LadderState::new(Direction::Up, Decimal::from(5), 5, chrono::Utc::now())
+        let ladder = LadderState::new(Direction::Up, 5, 5, chrono::Utc::now())
             .with_window_minutes(15);
         let ev = TraderEvent {
             ts: chrono::Utc::now(),
